worker/requirements.txt
eventlet==0.25.1
loguru==0.4.0
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.2
Flask-Login==0.4.1
Flask-Cors==3.0.8
sqlalchemy-utils==0.36.0
paramiko==2.7.1
textfsm==1.1.0
scp==0.13.2
six==1.13.0
pysnmp==4.4.9
pyserial==3.4
psycopg2==2.8.4
PyJWT==1.7.1
celery==4.3.0
redis==3.3.11
gunicorn==20.0.4
Flask-RESTful==0.3.7
pandas==0.25.3
requests==2.22.0
APScheduler==3.6.1
openpyxl==2.6.3
xlrd==1.2.0
